+++ CHANGELOG.md
@@ -56,6 +56,9 @@
 - [11073](https://github.com/vegaprotocol/vega/issues/11073) - Handle properly price factor < 1.
 - [11047](https://github.com/vegaprotocol/vega/issues/11047) - Add missing migration for recreating stop order views after the new fields were added.
 - [11090](https://github.com/vegaprotocol/vega/issues/11090) - Games API should only use the current team members.
+- [11089](https://github.com/vegaprotocol/vega/issues/11089) - Add tests for proto `enums` persisted to database.
+- []
+
 
 ## 0.75.0